refactor: rename Attributes to Annotations
